PER-66 Added deletion confirmation message to questions, categories, and options.
question.service.ts : Added a snackbar call to deleteCategory(), deleteMatch(), and deleteOption() that displays after the function completes. Created deleteQuestionById() under Question Functions, as a delete function didn't exist yet for questions. The function does not work - it only displays the snackbar message that a question has been deleted; I felt functionality was outside of the scope of the task, but can attempt it if I am incorrect.

list-questions.component.html : Attached the new deleteQuestionById() function to the delete button.