PER-123 PER-131 Added styling to the clickable download buttons. Added styling to the dialog box popup. Currently, the dialog box popup causes the page to shift to the right - this is something I'm going to have to research and will make a bug task for in Jira.
PER-123

styles.css : Added the 'downloadButton' class to share the styles of the upload button. Added a text color declaration which somehow wasn't necessary prior to this.

list-questions.component.html : Added 'downloadButton' style to the download button, and added a download icon to the display. Adjusted the layout of the code for better readability.

attachment.component.html : Added an extra space to the bottom of the component so it wouldn't sit directly next to the element below.

PER-131

confirmation-dialog.component.css : Added various specific styling elements.
confirmation-dialog.component.html : Added classes to the divs containing the message and the buttons.