PER-139 Added styling to validation error messages and related inputs.
custom-theme.scss : Changed the color of the Warn Base palette to red from Mid-State's yellow. This color value is used for the validation error messages, and the red color is more appropriate.

styles.css : Added 'error' class styling for use on error messages.

create-category.component.css : Added styling to the closeAddOption class to change the background color to Mid-State's yellow. This element previously utilized the "warn" color palette, and needed this addtion to retain it's yellow background color.
create-category.component.html : Removed the color definition on the closeAddOption button.

validation-messages.component.html : Added the 'error' class to the mat-error element.